requirements.txt
opencv-python
mediapipe
tensorflow
torch
transformers
sympy
fastapi
uvicorn
streamlit
onnxruntime
numpy